refactor: Numbers 클래스를 상속하는 TargetNumbers 클래스 구현
