+++ src/main/java/domain/Numbers.java
@@ -1,22 +1,18 @@
 package domain;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import camp.nextstep.edu.missionutils.Randoms;
-
 public class Numbers {
 
-	private static final int NUMBERS_SIZE = 3;
-	private static final int MIN_NUMBER = 1;
-	private static final int MAX_NUMBER = 9;
+	protected static final int NUMBERS_SIZE = 3;
+	protected static final int MIN_NUMBER = 1;
+	protected static final int MAX_NUMBER = 9;
 
-	private final List<Number> numbers;
+	protected List<Number> numbers;
 
 	public Numbers() {
-		this.numbers = generateRandomNumbers();
 	}
 
 	public Numbers(String input) {
@@ -40,19 +36,6 @@ public class Numbers {
 		return Arrays.stream(input.split(""))
 			.map(Number::new)
 			.collect(Collectors.toList());
-	}
-
-	private List<Number> generateRandomNumbers() {
-		List<Number> generatedNumbers = new ArrayList<Number>();
-
-		while (generatedNumbers.size() < NUMBERS_SIZE) {
-			Number nextNumber = new Number(Randoms.pickNumberInRange(MIN_NUMBER, MAX_NUMBER));
-			if (!generatedNumbers.contains(nextNumber)) {
-				generatedNumbers.add(nextNumber);
-			}
-		}
-
-		return generatedNumbers;
 	}
 
 }
