+++ src/main/java/domain/TargetNumbers.java
@@ -1,0 +1,26 @@
+package domain;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import camp.nextstep.edu.missionutils.Randoms;
+
+public class TargetNumbers extends Numbers {
+	public TargetNumbers() {
+		this.numbers = generateRandomNumbers();
+	}
+
+	private List<Number> generateRandomNumbers() {
+		List<Number> generatedNumbers = new ArrayList<Number>();
+
+		while (generatedNumbers.size() < NUMBERS_SIZE) {
+			Number nextNumber = new Number(Randoms.pickNumberInRange(MIN_NUMBER, MAX_NUMBER));
+			if (!generatedNumbers.contains(nextNumber)) {
+				generatedNumbers.add(nextNumber);
+			}
+		}
+
+		return generatedNumbers;
+	}
+
+}
